Refactor header and home components, add utils
Refactored the Header component to accept an activePage prop and highlight the active navigation link. Moved carousel, IT services overview, and services section components into a home subdirectory. Deleted the WhoWeAreSection component from about-us. Added a cn utility function for class name merging using clsx and tailwind-merge. Updated dependencies to include clsx and tailwind-merge.
