+++ src/components/home/carousel.tsx
